test: add service modules and coverage
